requirements.txt
pretty_midi
music21
customtkinter
pyyaml
pygame
